% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gerbil.R
\name{print.gof_gerbil}
\alias{print.gof_gerbil}
\title{Prints a \code{gof_gerbil} object. Printed output pertains to the goodness-of-fit tests that are applied in order to compare the distribution between observed and imputed cases for relevant variables or variable pairs.}
\usage{
\method{print}{gof_gerbil}(x, ...)
}
\arguments{
\item{x}{object of \code{gof_gerbil} class}

\item{...}{additional parameters to be passed down to inner functions.}
}
\value{
The functions \code{print.gof_gerbil} and \code{summary.gof_gerbil} display information 
  about the \code{cor_gerbil} object. The output displayed includes: 
  1) the average test statistic value across all variables or variable pairs contained in the object,
  2) the average p-value of all goodness-of-fit tests contained within the object, and
  3) the number of tests that yieled a p-value of less than 0.05.
}
\description{
Prints a \code{gof_gerbil} object. Printed output pertains to the goodness-of-fit tests that are applied in order to compare the distribution between observed and imputed cases for relevant variables or variable pairs.
}
