% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gerbil.R
\name{summary.cor_gerbil}
\alias{summary.cor_gerbil}
\title{Summarises a \code{gerbil} object. Printed output includes a variable-by-variable summary of variable types and missingness rates. The implemented predictor matrix is also provided.}
\usage{
\method{summary}{cor_gerbil}(object, ...)
}
\arguments{
\item{object}{An object of \code{cor_gerbil} class}

\item{...}{additional parameters to be passed down to inner functions.}
}
\value{
The functions \code{print.cor_gerbil} and \code{summary.cor_gerbil} display information 
  about the \code{cor_gerbil} object. The output displayed includes: 
  1) the average absolute difference in correlation between observed and imputed cases across all relevant variable pairs,
  2) the average value of the test statistic based on Fisher's z across all variable pairs,
  3) the largest test statistic observed across any variable pair, and
  4) the portion of p-values for the test based on Fisher's z that are less than 0.05.
}
\description{
Summarises a \code{gerbil} object. Printed output includes a variable-by-variable summary of variable types and missingness rates. The implemented predictor matrix is also provided.
}
