% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.gerbil}
\alias{summary.gerbil}
\title{Summarises a \code{gerbil} object. Printed output includes a variable-by-variable summary of variable types and missingness rates. The implemented predictor matrix is also provided.}
\usage{
\method{summary}{gerbil}(object, ...)
}
\arguments{
\item{object}{An object of \code{gerbil} class}

\item{...}{Additional parameters to be passed down to inner functions.}
}
\value{
The functions \code{print.gerbil} and \code{summary.gerbil} display information about the \code{gerbil} object.
  Primarily, the variable type and missingness rate are displayed for each variable. The predictor matrix is also provided.
}
\description{
Summarises a \code{gerbil} object. Printed output includes a variable-by-variable summary of variable types and missingness rates. The implemented predictor matrix is also provided.
}
